apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
langgraph>=0.2.74
google-genai>=1.0.0
langchain-core>=0.3.0
httpx>=0.27.0
